REFACTOR/ProductService - 수량 값 타입 변경

+++ product-service/src/main/java/com/palja/product_service/application/command/CreateProductCommand.java
@@ -6,7 +6,7 @@ public record CreateProductCommand(
         String name,
         String description,
         Long price,
-        Integer stock,
+        Long stock,
         String category
 ) {
