+++ product-service/src/main/java/com/palja/product_service/domain/repository/ProductRepository.java
@@ -30,15 +30,15 @@ public interface ProductRepository {
 
     ProductInfoForOrderDto findProductForOrder(UUID productId);
 
-    boolean decreaseStockBySale(String productId, Integer stock, Integer quantity);
+    boolean decreaseStockBySale(String productId, Long stock, Long quantity);
 
-    boolean adjustStock(String productId, Integer quantity);
+    boolean adjustStock(String productId, Long quantity);
 
     void stockBulkUpdateForSchedule(Collection<StockScheduleDto> dtos);
 
     boolean deleteStockFromRedis(String productId);
 
-    Product findByIdFetchStockWithLock(UUID productId, Integer quantity);
+    Product findByIdFetchStockWithLock(UUID productId, Long quantity);
 
     List<UUID> findAllIdsByCompanyUserId(UUID companyUserId);
